# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Admin:
    properties:
      avatar:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  entity.SuperAdmin:
    properties:
      avatar:
        type: string
      id:
        type: string
      is_blocked:
        type: boolean
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.Admin:
    properties:
      avatar:
        type: string
      id:
        type: string
      password:
        example: test
        type: string
      username:
        example: test
        type: string
    type: object
  models.AdminLoginRequest:
    properties:
      avatar:
        type: string
      password:
        example: test
        type: string
      username:
        example: test
        type: string
    type: object
  models.AdminLoginResponse:
    properties:
      access_token:
        type: string
    type: object
  models.Message:
    properties:
      message:
        type: string
    type: object
...
  models.SuperAdmin:
    properties:
      avatar:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.SuperAdminLoginRequest:
    properties:
      avatar:
        type: string
      password:
        example: test
        type: string
      phone_number:
        example: test
        type: string
    type: object
...
  v1.response:
    properties:
      error:
        example: message
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 'Backend - Nodirbek No''monov     TG: https://t.me/alwaysgolang'
  title: news back-end
  version: "1.0"
paths:
...
  /auth/admin/{id}:
    get:
      consumes:
      - application/json
      description: This method is getting admin by its id
      operationId: get-admin
      parameters:
      - description: ID of the admin to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Get Admin Data
      tags:
      - admin
  /auth/admin/create:
    post:
      consumes:
      - application/json
      description: Creates an admin
      operationId: superadmin-create-admin
      parameters:
      - description: Phone Number and Password to create Admin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Create Admin
      tags:
      - superadmin
  /auth/admin/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This method deletes admin.
      operationId: superadmin-delete-admin
      parameters:
      - description: ID of the admin to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Delete Admin
      tags:
      - superadmin
  /auth/admin/edit:
    put:
      consumes:
      - application/json
      description: ID of the admin to update and other fields will be updated.
      operationId: edit-admins
      parameters:
      - description: ID of the admin to edit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Edit Admin
      tags:
      - superadmin
  /auth/admin/getall:
    get:
      consumes:
      - application/json
      description: Gets All Admins
      operationId: get-all-admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Admin'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Get All Admins
      tags:
      - superadmin
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates an admin and returns an access token on successful
        login.
      operationId: admin-login
      parameters:
      - description: Phone Number and Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Login
      tags:
      - admin
  /auth/superadmin/edit:
    put:
      consumes:
      - application/json
      description: Superadmin updates by its id.
      operationId: edit-super-admin
      parameters:
      - description: Superadmin data for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SuperAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Edit Super Admin
      tags:
      - superadmin
  /auth/superadmin/login:
    post:
      consumes:
      - application/json
      description: Authenticates a super admin and returns an access token on successful
        login.
      operationId: superadmin-login
      parameters:
      - description: Phone Number and Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SuperAdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuperAdmin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Super Admin Login
      tags:
      - superadmin
...
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: API for file upload including images and videos
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Bucket type to put file
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: File upload
      tags:
      - file-upload
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"